Performance improvements to BinaryMinMaxHeap
The primary change is is removing allocations from the
`children_and_grandchildren` function. Originally, this function
allocated an array and pushed inbounds children and grandchildren to
this array. Since there are always six children and grandchildren (and
they are easy to compute), we can eagerly compute this collection and
return a lazy `Iterators.filter` for bounds checking.

Secondary changes including adding `@inbounds` and
`Base.@propagate_inbounds` and using array views where applicable.